add prsist and devtools
