+++ package.json
@@ -14,6 +14,7 @@
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
+    "redux-persist": "^6.0.0",
     "reselect": "^4.1.8",
     "sass": "^1.61.0",
     "styled-components": "^5.3.9",
