+++ src/store/store.js
@@ -1,24 +1,34 @@
 import { compose, applyMiddleware, createStore } from "redux";
 import { createLogger } from "redux-logger";
-
+import { persistStore, persistReducer } from "redux-persist";
+import storage from "redux-persist/lib/storage";
 import { rootReducer } from "./root-reducer";
 const logger = createLogger();
 
-const middleware = [logger];
-
-const composedEnhancers = compose(applyMiddleware(...middleware));
-
-export const store = createStore(rootReducer, undefined, composedEnhancers);
-
-const loggerMiddleware = (store) => (next) => (action) => {
-  if (!action.type) {
-    return next(action);
-  }
-  console.log("type:", action.type);
-  console.log("payload:", action.payload);
-  console.log("current state:", store.getState());
-
-  next(action);
-
-  console.log("next state:", store.getState());
+const persistConfig = {
+  key: "root",
+  storage,
+  blacklist: ["user"],
 };
+
+const persistedReducer = persistReducer(persistConfig, rootReducer);
+
+const middleware = [process.env.NODE_ENV !== "production" && logger].filter(
+  Boolean
+);
+
+const composeEnhancer =
+  (process.env.NODE_ENV !== "production" &&
+    window &&
+    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+  compose;
+
+const composedEnhancers = composeEnhancer(applyMiddleware(...middleware));
+
+export const store = createStore(
+  persistedReducer,
+  undefined,
+  composedEnhancers
+);
+
+export const persistor = persistStore(store);
